examples(tf-mnist): add train/attack demo with worker containers
Provide a full example that demonstrates how to integrate the tensorflow work containers with the
MLFlow Tracking service and Minio S3 backend. This example provides the scripts and code necessary
to train the LeNet-5 architecture on the MNIST dataset and then use the LeNet-5 weights to perform
the Fast Gradient Method evasion attack that generates adversarial images.