feat(docker): implement base training/attacking container
Install core set of Python libraries to support data science-related jobs. The primary machine
learning library provided in this container is scikit-learn. This image is to be used as a base for
deriving additional Docker images that support more specialized machine learning libraries,
particularly deep learning libraries.

By default, it is assumed that users will be leveraging the MLFlow Projects library to craft
workflows for their machine learning jobs.