feat(docker): implement image for serving the mlflow tracking server
Used to track machine learning experiments, log metadata, and provide lookup capabilities for the
MLFlow Model Registry. At startup, the service's entry point script looks up the S3 endpoint in an
environment variable that will be used to store artifacts. A bash script that parses this S3
endpoint URL will then check if the bucket it uses exists, and if it doesn't exist, create it prior
to launching the server.