feat(docker): implement restapi container
Install dependencies for a Flask-based RESTful API endpoint and also install the
current version of the project source code under the src directory. The endpoint in this image will
be the primary interface for users of the lab architecture. Users will send their instructions for
creating a job to the endpoint, and it will take care of the rest on its own.