examples(tf-mnist): add new architecture and entry-point parameters
Add a dense, shallow network architecture option to the model architecture registry. Add option to
name and register the model automatically after training is complete. Add option to set the random
number generator seed to use for reproducibility. Remove the explicit setting of intra- and inter-op
parallelism threads, as the tensorflow package in the anaconda channel will handle this
automatically.