feat(restapi): add timeout and dependency parameters to job endpoint
Add a "dependsOn" endpoint parameter to declare if a job should not run until another job has
finished. Leverages the one-to-one and one-to-many execution graphs implemented in the RQ Python
library. Add a "timeout" parameter to set how long a job should run before it is assumed to have
timed out. Assumes a default of 24 hours if this parameter is left blank.