feat(restapi): implement posting of workflow archive to job endpoint
Refactor the POST method of the job endpoint to expect requests as the `multipart/form-data` content
type. Add a field where users need to upload a workflow archive, which is then uploaded to a
temporary location on the S3 storage backend. Attach the URI to the Job entity in the database so
that it gets used during submission to the Redis-based job queue.