fix(docker): resolve boto3 crashes by replacing uwsgi with gunicorn
The version of uwsgi installed from conda-forge crashed whenever a call was made to the S3 storage
backend using the `boto3` package, rendering the POST method to the job endpoint on the API
unusable. Switching to gunicorn solved the problem.

Added an option to the entrypoints to toggle between the uwsgi and gunicorn servers. Extended the
uwsgi configuration while attempting to debug the issue, will keep it for now in case uwsgi becomes
usable again. Configured nginx to handle redirects to a gunicorn-based server instead of uwsgi.
Moved the `wsgi.py` file to the root of the project and updated file to choose the flask environment
using the `AI_RESTAPI_ENV` variable. Added option to trigger database migration upgrades to the
mlflow-tracking and restapi Docker image entrypoints.