feat(restapi): implement experiment endpoint and error handlers
Add a new endpoint for registering experiments that correspond to experiments in the MLFlow Tracking
server that are used to set experiment names at job runtime. Add a relationship connecting the
experiments and jobs tables by revising the Job model. Update the job endpoint to accept the
experiment name on the job submission form and to handle it in the endpoint controller and service.
Add and update unit tests for the new experiment endpoint functionality. Update the Docker image
MLFlow execution script to accept the experiment id as a parameter. Remove explicit setting of
experiment name using environment variables in tensorflow mnist example.

Added error handlers that return human-readable error messages to clients using the API. These are
registered and initialized at runtime.

Added a models.py file at the root of the restapi module, which eas necessary for resolving a
dependency importing issue for SQLAlchemy models. Refactored code to import from this file as
is necessary to avoid circular dependency conflicts.