examples(tf-mnist): split fgm entrypoint to create infer entrypoint
Extract the model evaluation step in the fgm entrypoint and move it into its own separate entrypoint
named infer. In the infer entrypoint, a user can point to a successful fgm run and use the
adversarial images generated there to run predictive inference against them using a model (either
the same one or a different one) and generate the typical evaluation metrics. This allows the demo
to showcase how the lab architecture facilitates examinations of attack transferability.