feat: implement queue endpoint and replace enums with constraints
Adds a new queue endpoint to the RESTful API for managing the list of queues watched by the worker
containers. Replaces all uses of the `Enum` data type (`JobStatus` and `JobQueue`) with strings that
are validated against uniqueness and key constraints that are enforced at the database level.
Provides migration scripts for the updated database schema.

There are two primary motivations for this change:

1. It is difficult to perform a database migration when expanding valid entries for an enum-typed
   column in a database

2. Adding and removing queues requires changing code and redeploying the endpoint, leading to less
   flexibility during runtime

Closes #30