refactor(docker): update and remove unused bash scripts
Remove old bash shell scripts that are not invoked in any of the current docker images and any
functions that were used to execute them. Update the copyright year to 2021 and update the default
conda environment to `mitre-securing-ai` in the remaining shell scripts. Update the argbash help
messages as necessary.