docs: restructure the documentation outline and add content
Refactor and restructure the table of contents for the documentation to be more user-focused
overall and migrate all existing documentation into the new structure. Add directives to
auto-generate the source code documentation for the task plugins and new modules in the SDK. Add
example docstrings to one of the artifacts task plugins. Integrate CONTRIBUTING.md into the
generated documentation.