feat(plugins): generalize demo functions into task plugins
This update creates the first official set of task plugins for the Securing AI Testbed, which are
installed by uploading them to the directory `securingai_builtins` of a S3 bucket that has been
designated for storing plugins. Many of these task plugins are generalizations of the functions used
in the train, fgm, and infer entry points of the tensorflow-mnist-classifier example. The following
changes made when generalizing the plugins are noted:

- All task plugins use the new `structlog` 20.2.0 type annotations
- Any functions that distinguish between gray-scale and color images now infer the color mode from
  the image size tuple instead needing it specified directly.
- The underlying implementation of the categorical accuracy metric has been revised