docs: add developers page and modify footer content
Add a page listing the names of developers that worked on the Securing AI Testbed that is linked at
the bottom of the navbar. Modified the footers in the main page and navbar to remove Copyright
messages and to show additional links in the navbar footer when files named `contact.rst` and
`terms.rst` are present.