chore: remove copyright and license statements and notices
Removes the LICENSE file along with all references in the configuration and metadata files, example
files, docker build files and shell scripts, and the source file headers to "The MITRE Corporation"
as the author and copyright holder. The author fields in the `setup.cfg` and PyPlugs `__init__.py`
file have been updated with the names of all of the development team members.

This is done in preparation for the upcoming source code transfer to NIST.