docs: first pass to replace internal project name with dioptra
The official name for this project is now Dioptra. This update is a first pass at replacing
references to "Securing AI Lab" and "Securing AI Testbed" in the documentation with "Dioptra".

References that could cause source code breakage if changed, such as the `mitre.securingai` import
path, will be migrated to the new name in a future update.