build(docker): upgrade critical dependencies and create legacy images
Migrate to Python 3.9 on all images and create legacy Tensorflow 2.1.0 and MLFlow Tracking 1.12.1
images that remain on Python 3.7 to ease transition to newer versions of Python, Tensorflow, and
MLFlow. Remove unnecessary package pinning in the conda environment files and in the Dockerfiles.
Update the build script to allow for cache overrides