build(docker): update dependencies in tensorflow and pytorch images
Switches from using conda to pip for installing the Tensorflow and PyTorch libraries to allow for
the latest versions of these packages to be installed. This requires changing the base image from
ubuntu to nvidia/cuda for the GPU-enabled images. In addition to upgrading to Tensorflow 2.8.0 and
PyTorch 1.10.1, a few additional packages have been added to the worker containers:

- imgaug
- pycocotools
- detectron2 (PyTorch only)

These libraries are provided to enable the use of object detection models within Dioptra in a future
update.

In addition, other dependencies have been bumped or repinned in the container images.