fix(restapi): declare explicit injector providers and enforce pep 563
Fixes the bug that appeared after the 3.15.0 update to the marshmallow package, specifically the
changes introduced in marshmallow-code/marshmallow#1932. These changes made the marshmallow package
compliant with PEP 563 and PEP 604, which introduced the `|` operator into the type annotations for
the constructor of the `Schema` class. The `typing.get_type_hints` function did not understand how
to evaluate the `|` operator in this context. The injector package uses `typing.get_type_hints` and
so it raised an error when it tried to check if the `Schema` class needed any injections.

This commit resolves this issue by declaring explicit injector providers for all classes inheriting
from marshmallow's `Schema` class that are known to be injected elsewhere in the code base. This
change avoids the issue entirely by preventing the injector library from trying to instantiate these
classes on its own. A provider is also implemented for the `mlflow.tracking.MlflowClient` class as a
future-proofing measure.

See also: marshmallow-code/marshmallow#1932, alecthomas/injector#116

Closes #30