refactor: migrate the repository to dioptra namespace
BREAKING CHANGE: Removes all the remaining references to "Securing AI" and to "MITRE" and replaces
them with "Dioptra". The references to MITRE is an artifact of this repository's development history
that are no longer needed. MITRE transferred this code to NIST, and NIST open-sourced it to GitHub.
The following is a list of the major breaking changes:

- The dioptra package's import namespace is changed from `mitre.securingai` to `dioptra`. All
  existing builtin and custom task plugins have been updated to reflect this change.
- The developer conda virtual environment was renamed from `securing-ai-testbed` to `dioptra`.
- The task plugin namespaces are changed from `securingai_builtins` and `securingai_custom` to
  `dioptra_builtins` and `dioptra_custom`.
- The taskPlugin REST API endpoints now use `dioptra_builtins` and `dioptra_custom` instead of
  `securingai_builtins` and `securingai_custom`.
- The name of the non-root user in the containers has been changed from `mitre` to `dioptra`.
- The name of the virtual environment in the containers has been changed from `mitre-securing-ai` to
  `dioptra`.
- The environment variable prefix used in the containers and the `dioptra` Python package has been
  changed from `AI_` to `DIOPTRA_`.

The namespace refactor was also applied to the project documentation as well as variables and
function names within the project configuration files, build scripts, Makefiles, Jupyter notebooks,
and unit and integration tests. Deprecated references to Securing AI Testbed and GitLab were removed
from the documentation. In addition, general typos were fixed in the documentation.

Co-authored-by: Paul Rowe <prowe@mitre.org>