refactor: remove legacy containers and associated scripts and tests
The legacy containers were used to prototype a script for migrating MLflow Tracking artifacts when
upgrading from an old set of containers based on Tensorflow 2.1.0 and MLflow 1.12.1 to the newer
versions. This script would only ever be used on old containers using code with the
`mitre.securingai` namespace. Since the namespace refactor from `mitre.securingai` to `dioptra` is
complete, this script is no longer needed and all legacy containers and associated tests can be
removed.