ci(github): disable integration tests
The integration tests hang when they are executed on the GitHub Action runners. These tests, which
run entirely within the containers, run fine on other machines. It is not fully understood why the
integration tests are now having trouble running on GitHub Actions. Until this is resolved, the
integration tests will be run on separate machines.