examples: add script and files for the updated road signs dataset
This provides updated annotations and a script that downloads the Road Signs Dataset
(https://makeml.app/datasets/road-signs), unpacks the zip archive, and reorganizes the files into a
predefined train/test split. The script also prepends all the filenames with a 5-digit prefix (00001
through 00136) that groups the images into "tracks". The 00000 prefix is reserved for images that do
not belong to a track.

A track is a sequence of correlated images sampled from a video clip of the same physical,
real-world object. Because the images within each track are highly correlated, each track should
either be placed in the training set or the testing set as a group. Splitting the images from a
single track across the training and testing set will result in data leakage.