feat(restapi): implement user endpoint
Adds a new user endpoint to the REST API for creating user accounts in Dioptra. This is a minimal
implementation: accounts can be created, retrieved, and deleted, and that's it. The authentication
component where a user must be logged in before they can access resources in other enpoints will be
added in a future update.

In addition to the REST API implementation, this update also includes the following:

- Documentation for the new endpoint
- Unit tests over the functionality of the new endpoint.
- Updated package lists for the containers needed by the Dioptra REST API