feat!: implement a template-based setup tool for configuring dioptra
The new setup tool is provided as a Cookiecutter template that allows the user to generate the
scripts, configuration files, and Docker Compose files needed to run Dioptra on a single machine.
The template variables are used to customize the generated files to suit a user's environment and
takes care of most of the configuration work for them. The generated files include an easy-to-use
initialization script that prepares the named container volumes, which provides optional support for
installing extra CA certificates into the containers and enabling/disabling SSL/TLS for the NGINX
and Postgres services.

The Dioptra Dockerfiles have been reworked to support the new setup tool, including changes to the
the baked-in configuration files, scripts, and the behavior of the container entrypoints. The
Dockerfiles and build toolchain also now supports users that need to include extra CA certificates
during an image build. This allows users to build the images within a corporate environment that has
its own certificate authority, see docker/ca-certificates/README.md for more information. Please
note that these certificates will only be used during the image build and that the final images will
not include them.

The documentation has been updated with a quickstart and detailed walkthrough that explains how to
build the updated containers and use the new setup tool. The old instructions and installation
sections have been removed. An alert has been added to the example documentation to point out that
a new setup tool is available, and that the documentation will be updated shortly.

BREAKING CHANGE: The new Dioptra container images are not compatible with the docker-compose.yml
files that used to be provided in the examples/ folder. Custom docker-compose.yml files that were
based on the guidance from the the now-removed section of the documentation, "Single Machine
Deployment," are also obsolete.

Closes #83