feat(ui): scaffold vue frontend and integrate with build tools
This update bootstraps a Vue.js frontend in the src/frontend directory that uses the Vite build tool
that provides both a development server and a build command for bundling up the frontend code. The
current frontend is a simple placeholder consisting of a "Welcome to Dioptra" page that lists the
built-in task plugins and worker queues that have been configured in the Dioptra application.

The following additional changes were necessary to integrate the frontend with the Flask backend:

- The Flask-Cors plugin has been added as a dependency to the REST API backend to add support for
  Cross-Origin requests when running the development server.
- The template-based setup tool has been extended to download and build the frontend files when
  running the init-deployment.sh script.
- The NGINX service will now be configured to serve the frontend and backend API over the same port.
- The `--plugin-branch` option in the init-deployment.sh script has been renamed to --branch, with
  that option now controling the branch the Vue.js source files are to be downloaded from.

This commit also contains the following changes unrelated to integrating the backend and frontend:

- The top-level README was updated with simple instructions for how to create developer virtual
  environments.
- The usage of `cookiecutter._output_dir` has been replaced with the `$((BASEDIR))` hook in the
  template for the systemd/dioptra.service file.

Closes #87
Co-authored-by: James K. Glasbrenner <jglasbrenner@mitre.org>