feat: implement script that fetches and organizes example datasets
Implement working downloaders for the MNIST, Road Signs Detection, and Fruits 360 datasets. Explain
the directory structure used for multi-class image classification and object detection modalities in
the docstring at the top of the script. Move the road-sign annotations from the
tensorflow-roadsigns-yolo/data folder into the scripts/patches folder. Add an
examples-setup-requirements.txt file to the scripts/venvs folder that can be used to create a
virtual environment that is able to run the download_data.py script.

The script also contains an in-progress implementation of the ImageNet downloader. Attempting to run
the ImageNet downloader via the CLI will print an alert with an message that the downloader is not
currently available. Todo messages that capture what still needs to be done to complete the ImageNet
downloader have been left in the appropriate places. User-facing notes that state that the ImageNet
downloader is still under construction have also been added to the script.

Add documentation to examples/README.md that explains how to setup and use the script to download
the datasets.

Closes #158
Co-authored-by: Julian T Sexton <JTSEXTON@MITRE.ORG>