refactor: remove the miniconda dependency from the container images
Removes the miniconda dependency from the container images, replaces it with system Python, venv,
and pip-tools, and adds `requirements.txt` lockfiles for the containers that are generated via tox.
The shell scripts executed in the containers have been updated to reflect this change.

Note that the REST API passes a `--conda-env` option to the run-mlflow-job.sh script that is
executed by the worker containers. Fully removing that option would result in a breaking change for
users running older builds of the worker containers. To prevent this, the option is retained in the
script and marked as deprecated. It will be fully removed in a future update.

In addition, this refactor removes the need for most of the `--build-arg` options that were being
passed to `docker build`, and so the `Makefile` build rules and the GitHub Actions configurations
have been simplified.