feat(task_engine): implement a new declarative task engine for dioptra
Adds the new task engine for Dioptra that replaces Prefect and allows users to create experiments
using a declarative YAML file. The structure of the YAML files is validated using a JSON-Schema
schema. In addition, the task engine includes a type validation module that checks if the task
plugin inputs and outputs are consistent throughout the task graph set up in the YAML file.
Two commandline scripts are added for running the task engine (`run-experiment`) and validating
the YAML files (`validate-experiment`).

This update introduces unit tests for this module and adds documentation to the Dioptra User Guide
that explains how the declarative task engine works. The project lockfiles have also been
regenerated.

Closes #101