fix: platform specific prefixes in s3 uris and mlflow tests
Addresses issue where S3 key prefixes followed the platform's path conventions causing tests
that validate S3 URIs to fail. The fix is to normalize S3 prefixes to follow POSIX path convention
using Path.as_posix() from pathlib.

Adds a GitHub Action for unit tests on a windows-latest image. test_mlflow.py was updated to use
pathlib instead of os.path to fix cross-platform issues with the tests. A fallback in the
cookiecutter template's post generation script was added for generating random letters as "words"
when a dictionary file is not present. The file encoding is explicitly specified in the cookiecutter
test.

Closes #248
Co-authored-by: James K. Glasbrenner <jglasbrenner@mitre.org>