build: migrate init-minio.sh to use new "mc admin policy" commands
Fixes the breaking changes that were introduced in the MinIO mc client earlier this year (2023) that
broke the init-minio.sh script provided in Dioptra's cookiecutter template. The init-minio.sh script
has been updated to use the new "mc admin policy" commands and handle the fact that the command for
attaching policies is no longer idempotent. The image tag for minio/mc has been unpinned and set
back to the "latest" tag in the dev branch.

Closes #207
Co-authored-by: James K. Glasbrenner <jglasbrenner@mitre.org>