examples: ensure scikit-learn is imported before tensorflow
There is an issue with glibc on Linux arm64 that causes dlopen errors when certain Python libraries
are loaded in the "wrong" order. This seems to happen when some packages, such as scikit-learn,
bundle their own copy of the libgomp library, but others do not. When this happens, the package
import order matters, as noted in the following GitHub Issue threads:

- https://github.com/opencv/opencv/issues/14884
- https://github.com/keras-team/keras-tuner/issues/317

The problematic libraries appear to be scikit-learn and Tensorflow, and the workaround is to make
sure that `import sklearn` comes before `import tensorflow`. To this end, this commit updates the
Python scripts for the Tensorflow examples to import scikit-learn early on, which ensures that the
import order is correct.