refactor: migrate to a cruft-managed cookiecutter deployment template
This update refactors the cookiecutter deployment template so that it can be managed using the
Cruft tool. This improvement allows users to sync updates to the deployment configuration without
needing to re-enter answers to the configuration questions and keeps the same auto-generated
passwords. This was made possible with the following changes:

- The auto-generated deployment passwords are now created when the user first invokes
  `init-deployment.sh`. The passwords are stored in a `.env` file that will automatically be read
  when the user invokes the `docker compose` command. The `docker-compose.yml` file was updated to
  accomodate this change.
- The documentation for configuring the deployment files now advises users to create a
  `docker-compose.override.yml` file to make changes to the deployment configuration. A sample
  template of the override file is rendered as part of the template.

The documentation has been updated to explain how to use Cruft to manage and update the deployment
files, and points out how the override file can even be used to add a custom worker container to the
deployment. In addition, the documentation has been updated to be more consistent, such as
consistenly referring to the path `/dioptra/data` as the container mount point for the datasets
directory.

Closes #126

Co-authored-by: Keith Manville <kmanville@mitre.org>
Co-authored-by: bjpatrick <137509145+bjpatrick@users.noreply.github.com>