feat: add pytest flag to run v1 REST API tests
This feature adds a --run-v1 flag to pytest in `tests/conftest.py` to enable
running of v1 API tests. To mark a test as v1, annotate the test function
with `@pytest.mark.v1`.

We intend to use this feature during the development of v1 of the API.
We expect tests to fail while the service layer is not implemented so v1
tests should be excluded by default.

To run only v1 tests use the command below:
````
python -m tox run -e py39-pytest -- tests/unit/restapi/v1 --run-v1
```

A copy of the v0 Queue tests has been placed in the v1 folder with the
API_ROOT set to v1 and all tests marked as v1 as a demonstration of this
feature. These tests to not properly exercise the new API and are
expected to fail.