feat: add pytest flag to run v1 REST API tests
This feature adds a `--run-v1` flag to pytest in `tests/conftest.py` to
enable running of v1 API tests. To mark a test as v1, annotate the test
function with `@pytest.mark.v1`.

We intend to use this feature during the development of v1 of the API.
Tests are expected to fail while the service layer is not implemented so
they should be excluded by default.

v1 tests should be placed in the `restapi/v1` directory.

To run only v1 tests use the command below:
```
python -m tox run -e py39-pytest -- tests/unit/restapi/v1 --run-v1
```