feat: proof-of-concept for re-using controller layer Endpoints
This is a proof-of-concept for re-using Endpoints from the controller
layer to achieve common functionality across multiple root Endpoints
with minimal code re-use.

The use case illustrated in this proof-of-concept is providing tagging
functionality via a `/<resource>/{id}/tags` endpoint for all Dioptra
Resources in the API. An Endpoint is created in `tags/controller.py`
which defines the operations for tagging resource, but it is not given a
route. In the other controller layers that wish to use this tagging
functionality, a subclass of the Endpoint is created and the `route`
decorator is applied.

Some customization can still be achieved via the constructor. For
example, we can set a `resource_name` member variable can be set to
customize logging messages. The correct Service object for the given
resource can be provided here. We might want some kind of base
ResourceService class that can handle common service layer operations
like tagging.