ci: disable the container-test workflow
This workflow is running a test that is flaky and also no longer needed for our purposes, since it
is just testing that the schema for the MLflow Tracking server has remained consistent between
upgrades. This commit prevents the test from running, and the testing code itself will be removed
in a future update.