build: migrate project toolchain to python 3.11 and ubuntu 22.04
This commit upgrades all of the project configuration files, GitHub Action workflows, and
Dockerfiles to use Python 3.11 and Ubuntu 22.04. The Dockerfiles use the deadsnakes PPA to install a
recent version of Python 3.11, since the Ubuntu package repository only provides the first release
candidate for Python 3.11.0, which is missing bug fixes. References to Python 3.9 have been replaced
with Python 3.11 in the project documentation.

This commit also upgrades the pinned version of Tensorflow to 2.16.1 and PyTorch to 2.2.2, which
both have installable wheel support for Python 3.11. This change required updating some of the
Dioptra plugins to work with the default eager execution model and the current API for Tensorflow.
The detectron2 and pytorch-lightning packages have been dropped due to lack of support and
compatibility with these versions of PyTorch and Python.

This commit also adds async_timeout as an explicit dependency for the worker and restapi images.
This fixes the problem introduced in https://github.com/redis/redis-py/pull/2602 where async_timeout
was removed as a dependency for Python 3.11, which had a ripple effect that resulted in unexpected
breakage. Adding the dependency back fixes the issue.

Co-authored-by: Keith Manville <kmanville@mitre.org>