fix: update v1 schema ID fields to drop 'Id' suffix
This fix uses marshmallow's data_key parameter to rename ID fields so
that they don't have the 'Id' suffix. The 'Id' suffix was previosuly
used when a schema was used in both response and request, but the type
for the field was different (a RefSchema or an integer ID). The data_key
parameter allows the same field name to be used in both response and
request.