feat: implement the v1 database schema and sqlalchemy orm classes
This update creates the database tables and ORM classes for the version 1 release of Dioptra in
accordance with the design captured in the project's Entity-Relationship diagram. This redesign
provides historical lineage tracking and strong data integrity guarantees for running and tracking
machine learning experiments using Dioptra's plugin and task engine systems. It is also designed to
support and store the state of user interactions in the new Dioptra frontend.

This implementation uses the new dataclass integrations provided in SQLAlchemy 2.0. The legacy ORM
models have also been updated to use SQLAlchemy 2.0 features for consistency with the v1 models.
SQLite database connections are also automatically configured to enforce foreign key constraints. In
addition, the following custom data types have been added to support the v1 implementation:

- A backend-agnostic timezone-aware datetime type called `TZDateTime`. This type automatically
  converts incoming timestamps to UTC before storing them in the database.
- Multiple type aliases for common data types used throughout the schema, such as `intpk` for
  integer primary keys and `str255` for `String(255)` text fields.

The ORM classes make extensive use of SQLAlchemy's features to make it easier to work with related
data objects. SQLAlchemy relationships and table inheritance are used where possible to make
book-keeping easier. The column_property feature is also used to create derived columns based on
data in other tables, such as an `is_deleted` field that is based on the existance of a "delete"
lock.

New alembic migration scripts have been generated as part of this update, which include
Postgres-specific fixes to ensure there are no naming conflicts between the v0 and v1 tables,
sequences, and constraints. The migration script also sets up the v1 ontology by inserting records
into Enum-style tables.

A unit testing suite has been added to exercise the ORM classes and ensure that the schema is
functioning as expected. The tests leverage fixtures to set up test data in the database. A
`FakeData` class is used to generate fake data using the `Faker` library. The tests cover inserting
data into the v1 tables, querying the data using custom view queries, and verifying that the
relationships and constraints are working properly. There are also a couple of tests that show how
to use draft resources to store and retrieve an in-progress edit to a Dioptra entity and how to use
resource snapshots to create a history of "append-only" edits to an existing resource. It should be
noted that these tests are intended to be temporary and to serve as basic documentation and guidance
for how the ORM classes should be used in the service layer. As the service layer is built out and
it gets hooked up with the controller layer, these ORM tests should be replaced by the REST API
integration tests.

Additional changes include:

- The project lockfiles have been regenerated to bring in the `Faker` library for use in testing.
- The API documentation has been updated to remove defunct sections.
- The Mypy static type checker test now installs the Dioptra package and its dependencies in the Tox
  environment instead of skipping the install step.

Closes #464