feat: REST API v1 service layer for Queues
This feature adds an implementation of the service layer for v1 of the
REST API for the Queues endpoints.

The primary implementations are in `restapi/queues/service.py`. The
services are injected into the controller layer in
`restapi/queues/controller.py` which calls the appropriate service layer
functionality. The controller layer uses helper functions from
`restapi/utils.py` to translate ORM objects back to the response types
expected by the schema. Error handling is added to the v1 routes by
registering error handlers in `restapi/errors.py`.

The existing unit tests have been modified to correct errors and in some
cases expanded to catch additional edge cases.

It adds an additional derived field, `latest_snapshot_id` to the
`Resource` ORM class. This is used throughout the service layer to query
for the latest snapshot.