feat: implemented drafts for the v1 Queues endpoint
This feature implements drafts for the Queues endpoint. It includes the
drafts schema, controller, and service layers. The Draft controller
endpoints are generated by a function to customize them for the target
resource. The drafts services are customized via dependency injection so
separate service implementations are not needed for each resource. The
resource schemas are used to validate the contents of draft payloads.

Tests for drafting queues have been added to the queues test suite.