feat: REST API v1 service layer for Tags
This feature adds an implementation of the service layer for v1 of the REST
API for the Tags endpoint. It includes unit tests that cover common user
interactions with these resources. It also includes the tags subendpoint that
allows for the management of tags for the Resource endpoints.

The primary implementations are in `restapi/tags/service.py`. The services
are injected into controller layer in `restapi/tags/controller.py` which calls
the appropriate service layer functionality. The controller layer uses helper
functions from `restapi/utils.py` to translate ORM objects back to the response
types expected by the schema. Error handling is added to v1 routes by
registering error handlers in `restapi/errors.py`.

The implementation of the tags subendpoint is in `restapi/shared/tags`. The
tagging functionality has been integrated for the Queues endpoint. The test
suite for queues has been expanded to cover tagging operations.

Co-authored-by: Andrew Hand <ahand@mitre.org>
Co-authored-by: James K. Glasbrenner <jglasbrenner@mitre.org>